+++ NEWS.md
@@ -1,5 +1,10 @@
 # officer 0.6.3.9000
 
+## Features 
+
+- `fp_border()` gains support for all line border styles listed in ECMA-376 
+  § 17.18.2 and allowed CSS border styles. Closes #165 and #443.
+
 ## Changes
 
 - the dataframe returned by `docx_comments()` gains a list column `para_id` containing